[upd] - move login and register controller to auth folder
